Scrapy spider to get transition times
Ned hacked it up.  Scrapy is pretty good.  JIRA is crazy.

Output is states.json.  This isn't all the info we need, because the transitions
table doesn't include the latest state, so we need to do that in addition.
But at least it proves the concept of getting this information from JIRA.