feat: auto-create extras_require in setup.py
A few tools require burdensome third-party packages, and have been using separate requirements/*.txt files to manage
them. This makes it difficult to instruct people how to install them.

Now a directory can have an extra.in file, and it will automatically be included in `make upgrade`, and will
automatically become an extra that can be installed.