chore: Add auto pull version from tutor's config
Add functions for automatically pulling version number from tutor's
default config file and choosing which bars to bold and outline.

This pulls the name and / or number for the current version of openedx,
MongoDB, Elasticsearch, MySQL, and Redis. It does not pull the current
version of Python, Django, Ubuntu, Node, or Ruby.

To be clear, this commit doesn't add new bars to the calendar, it only
automates how we choose which bars to bold and outline. For example, if
we were to switch to Elasticsearch 8.x.x, we'll still need to manually
add the release and EOL dates.