fix: The annotations section can exist and be empty.
For example:
```
metadata:
    annotations: null
```

This would be valid but would throw an error before this change because
the code tries to use the `in` operator as if annotations is a
collection.  But if it's a NoneType, it throws an error since a NoneType
doesn't support the `in` operator.

We instead use the get method on the metadata collection which will
return `None` if the key is not in the dict.  This allows us to bypass
the `in` check for the `annotations` key and fix a potential error
scenario because the code behaves the same way for the key not existing
and for the value being none which is what we want in this case.