feat: Add a new version of the tag_release.py script.
Add a simpler alternative to the tag_release.py file.  The original
tag_release has grown and is hard to modify and has issues/errors.
Rather than trying to refactor it, here is a much simpler version that
should do all the work we normally do as a part of the release tagging
process.