fix: Compute diff properly in package home URL sets (#616)
Using symmetric difference caused entries from the ignore list to be
reported as 2nd party orgs in some cases. This appears to happen when the
lookup fails for a package that matches something in the ignore list, and
so its URL doesn't end up in `packages_urls`.

It's not clear what the nature of the failure is exactly (since we're not
getting log messages about non-200 status codes) but we were seeing
sporadic claims that `https://github.com/edx/codejail-includes` was found
in the dependencies. The package on PyPI now actually is under the openedx
org, so I believe this URL must have been coming from the ignore list.

We could add logging to discover what the failure mode is, but a lot of
packages don't have a home URL, so it may not be simple to do this in a way
that isn't noisy.